Put private api in separate directory
