+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shellicar/core-di",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "A basic dependency injection library",
   "repository": {
     "type": "git",
