+++ src/logger.ts
@@ -1,8 +1,6 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/no-unused-vars */
 export abstract class ILogger {
-  public debug(message?: any, ...optionalParams: any[]) {}
-  public info(message?: any, ...optionalParams: any[]) {}
-  public error(message?: any, ...optionalParams: any[]) {}
-  public warn(message?: any, ...optionalParams: any[]) {}
+  public debug(_message?: any, ..._optionalParams: any[]) { }
+  public info(_message?: any, ..._optionalParams: any[]) { }
+  public error(_message?: any, ..._optionalParams: any[]) { }
+  public warn(_message?: any, ..._optionalParams: any[]) { }
 }
